+++ contract-testing/src/constants.rs
@@ -75,3 +75,9 @@ pub(crate) const INITIALIZE_CONTRACT_INSTANCE_CREATE_COST: Energy = Energy {
 pub(crate) const UPDATE_CONTRACT_INSTANCE_BASE_COST: Energy = Energy {
     energy: 300,
 };
+
+/// The name of the environment variable that holds the path to the contract
+/// module file. To load the module, use the
+/// [`module_load_contract`](crate::module_load_contract) function.
+pub const CONTRACT_MODULE_OUTPUT_PATH_ENV_VAR: &'static str =
+    "CARGO_CONCORDIUM_TEST_MODULE_OUTPUT_PATH";